+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- `totalizers` prop in `Table` component.
+
 ## [9.101.0] - 2019-12-18
 
 ### Added